Collapse utils.
