+++ preprocessing/data_handling.py
@@ -1,6 +1,8 @@
 import os
 from random import Random
+from typing import Union
 
+import numpy as np
 import tensorflow as tf
 
 
@@ -62,3 +64,52 @@ def load_files(metafile,
                 break
         alphabet = sorted(list(alphabet))
         return samples, alphabet
+
+
+class Tokenizer:
+    
+    def __init__(self, alphabet, start_token='>', end_token='<', pad_token='/'):
+        self.alphabet = alphabet
+        self.idx_to_token = {i: s for i, s in enumerate(self.alphabet, start=1)}
+        self.idx_to_token[0] = pad_token
+        self.token_to_idx = {s: i for i, s in self.idx_to_token.items()}
+        self.start_token_index = len(self.alphabet) + 1
+        self.end_token_index = len(self.alphabet) + 2
+        self.vocab_size = len(self.alphabet) + 3
+        self.idx_to_token[self.start_token_index] = start_token
+        self.idx_to_token[self.end_token_index] = end_token
+    
+    def encode(self, sentence, add_start_end=True):
+        sequence = [self.token_to_idx[c] for c in sentence if c in self.token_to_idx]
+        if add_start_end:
+            sequence = [self.start_token_index] + sequence + [self.end_token_index]
+        return sequence
+    
+    def decode(self, sequence):
+        return ''.join([self.idx_to_token[int(t)] for t in sequence if int(t) in self.idx_to_token])
+
+
+class DataPrepper:
+    
+    def __init__(self,
+                 config,
+                 tokenizer: Union[Tokenizer]):
+        self.start_vec = np.ones((1, config['mel_channels'])) * config['mel_start_value']
+        self.end_vec = np.ones((1, config['mel_channels'])) * config['mel_end_value']
+        self.tokenizer = tokenizer
+        self.mel_channels = config['mel_channels']
+    
+    def __call__(self, sample, include_text=True):
+        phonemes, text, mel_path = sample
+        mel = np.load(mel_path)
+        return self._run(phonemes, text, mel, include_text=include_text)
+    
+    def _run(self, phonemes, text, mel, *, include_text):
+        encoded_phonemes = self.tokenizer.encode(phonemes)
+        norm_mel = np.concatenate([self.start_vec, mel, self.end_vec], axis=0)
+        stop_probs = np.ones((norm_mel.shape[0]))
+        stop_probs[-1] = 2
+        if include_text:
+            return norm_mel, encoded_phonemes, stop_probs, text
+        else:
+            return norm_mel, encoded_phonemes, stop_probs
