+++ utils/vec_ops.py
